Rewritten
Changed some methods around to make the flow of giving reputation more flashy.

When the "rep" button is clicked there is now no delay until the form is shown, which will make the user aware that them clicking the button has been recognized.

When a rating is added ("add rating" button is clicked) the page does no longer reload, instead the text "Updated!" is shown for 3.5 seconds.

Styling was change a bit. It now also uses the GM_getResourceText and GM_addStyle functions for handling styling. The CSS file is currently hosted on gist.github.com.